requirements.txt
XlsxWriter
Pillow